wired in real direct_submit; but test data has non-fits so exits

+++ tada/watch.py
@@ -2,18 +2,15 @@
 """When files appear in watched directory, ingest them. This should be
 wrapped into a service."""
 
-'''
-TODO:
-- YAML for personality files
-'''
-
 import argparse
 import logging
 import os
+import os.path
 import random
 import time
 import shutil
 from pathlib import PurePath
+from glob import glob
 
 import yaml
 
@@ -21,6 +18,29 @@ import watchdog.events
 import watchdog.observers
 
 from . import submit as ts
+from . import config
+
+'''
+Expected directory structures:
+/.../dropbox/
+  instrument1/...
+  instrument2/...
+  ...
+
+/.../valley-stash/
+  ingested/instrument1/...
+  rejected/instrument1/...
+
+To submit a batch (e.g. for pipeline or new instrument):
+  rsync -az ~/myfiles/newinstrument /.../dropbox/newinstrument
+
+If YAML files are found, they will be used for personalities.  For each new fits file,
+two locations will be looked in for YAML files: 
+  /.../dropbox/<instrument>/*.yaml
+  /.../dropbox/<instrument>/.../<terminal_dir>/*.yaml
+
+'''
+
 
 
 def fake_submit(ifname, checksum, qname, qcfg):
@@ -30,44 +50,69 @@ def fake_submit(ifname, checksum, qname, qcfg):
         raise Exception('FAILED')
     return ifname
 
-def new_file(ifname, watched_path, rejected_path, dest_path=None):
-    p = PurePath(ifname).relative_to(watched_path)
-    parts=p.parts
-    logging.debug('DBG: new_file: ifname={}, p={}, idir={}'
-                  .format(ifname, p.path, parts))
-    #if isinstance(event, watchdog.events.FileCreatedEvent):
-    if p.suffix == '.fz' or p.suffix == '.fits':
-        logging.debug('Got FITS: {}'.format(ifname))
-        destfname=(rejected_path / p).path
-        try:
-            #destfname = ts.submit_to_archive(ifname, checksum, 
-            #                                 qname, qcfg)
-            fake_submit(ifname, 'checksum', 'qname', 'qcfg')
-        except Exception as sex:
-            # FAILURE: stash it
-            logging.info('Ingest FAILED: stash into: {}'
-                          .format(destfname))
-            shutil.move(ifname, destfname)
-        else:
-            # SUCCESS: remove it
-            logging.info('Ingest SUCCEEDED: remove: {}'.format(ifname))
-            os.remove(ifname)
-    
 
 class SubmitEventHandler(watchdog.events.FileSystemEventHandler):
-    def __init__(self, watched_dir, rejected_dir):
+    def __init__(self, watched_dir, rejected_dir, moddir, qcfg):
         self.watched_path = PurePath(watched_dir)
         self.rejected_path = PurePath(rejected_dir)
+        self.moddir = moddir
+        self.qcfg = qcfg
         super(watchdog.events.FileSystemEventHandler).__init__()
         
     def on_created(self,event):
-        new_file(event.src_path, self.watched_path, self.rejected_path)
+        self.new_file(event.src_path)
     def on_moved(self,event):
         # for rsync: moved from tmp file to final filename
-        new_file(event.dest_path, self.watched_path, self.rejected_path)
+        self.new_file(event.dest_path)
+
+    def new_file(self, ifname):
+        p = PurePath(ifname).relative_to(self.watched_path)
+        parts=p.parts
+        logging.debug('DBG: new_file: ifname={},parts={}'.format(ifname, parts))
+
+        #if isinstance(event, watchdog.events.FileCreatedEvent):
+        if p.suffix == '.fz' or p.suffix == '.fits':
+            logging.debug('Got FITS: {}'.format(ifname))
+            pdir = self.watched_path / p.parts[0] / p.parts[1]
+            logging.debug('DBG: pdir={}'.format(pdir))
+
+            pdict = dict(options={}, params={})
+            for yfile in sorted(glob(str(pdir) + '/' + '*.yaml')):
+                logging.debug('DBG: reading YAML {}'.format(yfile))
+                with open(yfile) as yy:
+                    yd = yaml.safe_load(yy)
+                    pdict['params'].update(yd.get('params',{}))
+                    pdict['options'].update(yd.get('options',{}))
+
+            destfname = str(self.rejected_path / p)
+            yfile = ifname + '.yaml'
+            if os.path.isfile(yfile):
+                logging.debug('DBG: reading YAML {}'.format(yfile))
+                with open(yfile) as yy:
+                    yd = yaml.safe_load(yy)
+                    pdict['params'].update(yd.get('params',{}))
+                    pdict['options'].update(yd.get('options',{}))
+
+            logging.debug('DBG: pdict={}'.format(pdict))
+            try:
+                #!fake_submit(ifname, 'checksum', 'qname', 'qcfg')
+                destfname = ts.direct_submit(ifname, self.moddir,
+                                             personality=pdict,
+                                             qcfg=self.qcfg)
+            except Exception as sex:
+                # FAILURE: stash it
+                logging.info('Ingest FAILED: stash into: {}; {}'
+                              .format(destfname, sex))
+                os.makedirs(os.path.dirname(destfname), exist_ok=True)
+                shutil.move(ifname, destfname)
+            else:
+                # SUCCESS: remove it
+                logging.info('Ingest SUCCEEDED: remove: {}'.format(ifname))
+                os.remove(ifname)
+        
                 
-def ingest_drops(watched_dir, rejected_dir):
-    handler = SubmitEventHandler(watched_dir, rejected_dir)
+def ingest_drops(watched_dir, rejected_dir, moddir, qcfg):
+    handler = SubmitEventHandler(watched_dir, rejected_dir, moddir, qcfg)
     observer = watchdog.observers.Observer()
     logging.info('Watching directory: {}'.format(watched_dir))
     observer.schedule(handler, watched_dir, recursive=True)
@@ -87,11 +132,34 @@ def main():
         description='Ingest or stash files as the appear in watched directory',
         epilog='EXAMPLE: %(prog)s a b"'
         )
+    dflt_moddir = os.path.expanduser('~/.tada/submitted')
+    dflt_config = os.path.expanduser('~/.tada/config.json')
+    logconf = os.path.expanduser('~/.tada/logging.yaml')
     parser.add_argument('watched_dir',
                         help='Try to ingest every file dropped into this '
                         'directory (or its subdirectories)')
     parser.add_argument('rejected_dir',
                         help='Move files that fail to ingest to this directory')
+    parser.add_argument('-m', '--moddir',
+                        default=dflt_moddir,
+                        help="Directory that will contain the (possibly modified, possibly renamed) file as submitted. Deleted after iRODS put. [default={}]".format(dflt_moddir),
+                        )
+    parser.add_argument('--logconf',
+                        help='Logging configuration file (YAML format).'
+                        '[Default={}]'.format(logconf),
+                        default=logconf,
+                        type=argparse.FileType('r'))
+    parser.add_argument('-c', '--config',
+                        default=dflt_config,
+                        help='Config file. [default={}]'.format(dflt_config),
+                        )
+    parser.add_argument('-t', '--timeout',
+                        type=int,
+                        help='Seconds to wait for Archive to respond',
+                        )
+    parser.add_argument('--trace',
+                        action='store_true',
+                        help='Produce stack trace on error')
     parser.add_argument('--loglevel',      help='Kind of diagnostic output',
                         choices = ['CRTICAL','ERROR','WARNING','INFO','DEBUG'],
                         default='WARNING',
@@ -105,9 +173,14 @@ def main():
                         format='%(levelname)s %(message)s',
                         datefmt='%m-%d %H:%M'
                         )
-    logging.debug('Debug output is enabled!!!')
+    logDict = yaml.load(args.logconf)
+    logging.config.dictConfig(logDict)
+    logging.getLogger().setLevel(log_level)
 
 
-    ingest_drops(args.watched_dir, args.rejected_dir)
+    qcfg, dirs = config.get_config(None,
+                                   validate=False,
+                                   json_filename=args.config)
+    ingest_drops(args.watched_dir, args.rejected_dir, args.moddir, qcfg=qcfg)
 if __name__ == '__main__':
     main()
